Progress on functional rotation
