+++ src/main/java/com/boundless/mixin/flight_ability/PlayerFlightAccessRendering.java
@@ -1,6 +1,9 @@
 package com.boundless.mixin.flight_ability;
 
+import com.boundless.ability.components.KeybindHoldData;
 import com.boundless.ability.reusable_abilities.flight.FlightRendering;
+import com.boundless.hero.SuperHero;
+import com.boundless.registry.DataComponentRegistry;
 import com.boundless.util.FlightAccess;
 import com.boundless.util.HeroUtils;
 import net.minecraft.client.network.AbstractClientPlayerEntity;
@@ -9,11 +12,15 @@ import net.minecraft.client.render.entity.LivingEntityRenderer;
 import net.minecraft.client.render.entity.PlayerEntityRenderer;
 import net.minecraft.client.render.entity.model.PlayerEntityModel;
 import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.util.math.RotationAxis;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import java.util.HashMap;
+import java.util.Map;
 
 @Mixin(PlayerEntityRenderer.class)
 public abstract class PlayerFlightAccessRendering extends LivingEntityRenderer<AbstractClientPlayerEntity, PlayerEntityModel<AbstractClientPlayerEntity>> implements FlightAccess {
@@ -26,10 +33,21 @@ public abstract class PlayerFlightAccessRendering extends LivingEntityRenderer<A
 
     @Inject(at = @At(value = "HEAD"), method = "setupTransforms(Lnet/minecraft/client/network/AbstractClientPlayerEntity;Lnet/minecraft/client/util/math/MatrixStack;FFFF)V", cancellable = true)
     protected void boundless$flightTransforms(AbstractClientPlayerEntity abstractClientPlayerEntity, MatrixStack matrixStack, float f, float g, float tickDelta, float i, CallbackInfo ci) {
+        /*
         if (!abstractClientPlayerEntity.getAbilities().flying || !HeroUtils.isHero(abstractClientPlayerEntity)) return;
         super.setupTransforms(abstractClientPlayerEntity, matrixStack, f, g, tickDelta, i);
-        FlightRendering.hoverRendering(abstractClientPlayerEntity, matrixStack, f, g, tickDelta, i, (PlayerEntityRenderer)(Object)this, ci);
+        //FlightRendering.hoverRendering(abstractClientPlayerEntity, matrixStack, f, g, tickDelta, i, (PlayerEntityRenderer)(Object)this, ci);
+        if (!abstractClientPlayerEntity.isFallFlying()) {
+            FlightRendering.renderFlight(abstractClientPlayerEntity, matrixStack, f, g, tickDelta, i, (PlayerEntityRenderer)(Object)this);
+        }
         ci.cancel();
+
+         */
+        //super.setupTransforms(abstractClientPlayerEntity, matrixStack, f, g, tickDelta, i);
+        Map<String, KeybindHoldData> keyHeldMap = HeroUtils.getHeroStack(abstractClientPlayerEntity).getOrDefault(DataComponentRegistry.HELD_KEYBIND, new HashMap<String, KeybindHoldData>());
+        KeybindHoldData keybindHoldData = keyHeldMap.get("key.forward");
+        if (keybindHoldData == null) return;
+        matrixStack.multiply(RotationAxis.POSITIVE_X.rotationDegrees(keybindHoldData.startTimestamp()));
     }
 
     @Override
