+++ src/main/java/com/boundless/registry/CodecRegistry.java
@@ -5,13 +5,11 @@ import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 
 public class CodecRegistry {
-    public static final Codec<KeybindHoldData> KEYBIND_HOLD_CODEC = RecordCodecBuilder.create(builder -> {
-        return builder.group(
-                Codec.BOOL.fieldOf("held").forGetter(KeybindHoldData::held),
-                Codec.LONG.fieldOf("startTimestamp").forGetter(KeybindHoldData::startTimestamp),
-                Codec.LONG.fieldOf("endTimestamp").forGetter(KeybindHoldData::endTimestamp)
-        ).apply(builder, KeybindHoldData::new);
-    });
+    public static final Codec<KeybindHoldData> KEYBIND_HOLD_CODEC = RecordCodecBuilder.create(builder -> builder.group(
+            Codec.BOOL.fieldOf("held").forGetter(KeybindHoldData::held),
+            Codec.LONG.fieldOf("startTimestamp").forGetter(KeybindHoldData::startTimestamp),
+            Codec.LONG.fieldOf("endTimestamp").forGetter(KeybindHoldData::endTimestamp)
+    ).apply(builder, KeybindHoldData::new));
 
     public static void initialize() {}
 }
