+++ src/main/java/com/boundless/util/KeybindingUtils.java
@@ -19,27 +19,4 @@ public class KeybindingUtils {
         Map<String, KeyBinding> keysByID = KeybindAccessor.getKeysByID();
         return keysByID.get(translation);
     }
-
-    public static void inputLogic(MinecraftClient client, String translatableKey) {
-        if (KeybindingUtils.getKeyBindingFromTranslation(translatableKey).isPressed()) {
-            Identifier abilityID = AbilityUtils.abilityIDFromKeybind(client.player, translatableKey);
-            if (AbilityUtils.checkAndUseAbility(client.player, abilityID)) {
-                ClientPlayNetworking.send(new AbilityUsePayload(abilityID));
-            };
-        }
-    }
-
-    private static final Map<String, Boolean> heldKeysMap = new HashMap<>();
-
-    public static void keybindHoldLogic(MinecraftClient client, KeyBinding key, String translationKey) {
-        if (client.player == null) return;
-
-        if (key.isPressed() && !heldKeysMap.getOrDefault(translationKey, false)) {
-            heldKeysMap.put(translationKey, true);
-            ClientPlayNetworking.send(new UpdateHoldStatePayload(translationKey, true));
-        } else if (!key.isPressed() && heldKeysMap.getOrDefault(translationKey, false)) {
-            heldKeysMap.put(translationKey, false);
-            ClientPlayNetworking.send(new UpdateHoldStatePayload(translationKey, false));
-        }
-    }
 }
